Move depthai version check in depthai_helpers; check version in calibration too; address review comments
