+++ depthai_demo.py
@@ -15,30 +15,8 @@ import depthai
 print('Using depthai module from: ', depthai.__file__)
 print('Depthai version installed: ', depthai.__version__)
 
-def get_version_from_requirements():
-    with Path('requirements.txt').open() as f:
-        datafile = f.readlines()
-    for line in datafile:
-        if 'depthai' in line:
-            #not commented out
-            if(line.split()[0] != '#'):
-                try:
-                    version = line.split('==')[1]
-                    #remove any whitespace
-                    version = version.strip()
-                except:
-                    version = None
-                return version
-    return None
-
-version_required = get_version_from_requirements()
-if version_required is not None:
-    print('Depthai version required:  ', version_required)
-    if version_required != depthai.__version__:
-        raise ValueError(f"\033[1;5;31mVersion mismatch \033[0m\033[91m between installed depthai lib and the required one by the script.\033[0m \n\
-            Required:  {version_required}\n\
-            Installed: {depthai.__version__}\n\
-            \033[91mRun: ./install_requirements.sh \033[0m")
+from depthai_helpers.version_check import check_depthai_version
+check_depthai_version()
 
 import consts.resource_paths
 from depthai_helpers import utils
